// superset-frontend/src/features/productAssets/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

export type AssetType = '图表' | '仪表盘' | '数据集' | '数据库';

export type AssetStatus = '已发布' | '草稿' | '已归档';

export interface ProductAsset {
  id: number;
  name: string;
  type: AssetType;
  owner: string;
  department: string;
  createdOn: string;
  lastModified: string;
  status: AssetStatus;
  description: string;
  tags: string[];
}